requirements.txt
jira
pdfkit
beautifulsoup4
requests
openai
pandas
tqdm
scikit-learn
numpy
faiss-cpu
fastapi 
uvicorn
python-dotenv
matplotlib
seaborn
plotly